Fuzzy Logic System for Agoda
Target Application used for Fuzzy Logic is Hotel Ratings under E-commerce domain. Here, for preparing Fuzzy Logic considered Agoda Booking site as example. Once the 
destination and dates of travels are provided the site provides us a list of hotels based on customer ratings. The customer ratings are scaled on 0 to 10 based on certain criteria.