Classification & Regression ML-Azure ML Studio (designer)
For this task you need to design and deploy a machine learning model using Azure ML studio (designer). You need to use Microsoft Azure Machine Learning Studio to design and deploy your model. To complete this task, you need to select a dataset and design a decision tree model (classification tree or regression tree) and then deploy the built model and get the API key. To do this task you need to follow the workshop recording and slides and deploy your own model on Azure. You need to provide the screenshots of your designed model, training model, the performance of the built model (e.g., Accuracy, confusion matrix and etc) and deployed model with the API key and test the model.